ED-015: Changed value types to floats

+++ code/creature.h
@@ -13,6 +13,7 @@
  *      22OCT2021 ED-009: Removed iostream include and cout commands
  *      24OCT2021 ED-013: Added hit detection
  *      26OCT2021 ED-014: Added rotation handling
+ *      26OCT2021 ED-015: Change value types to floats
  **/
 
 #ifndef CREATURE_H
@@ -30,12 +31,12 @@ struct ScoreToken
 class Creature
 {
 public:
-    Creature(int xDimension, int yDimension, int x, int y):
+    Creature(float xDimension, float yDimension, float x, float y):
         xDim(xDimension),yDim(yDimension),xPos(x),yPos(y){setPos(x,y);}
-    Creature(int xDimension, int yDimension, int x, int y, std::string ID):
+    Creature(float xDimension, float yDimension, float x, float y, std::string ID):
         xDim(xDimension),yDim(yDimension),xPos(x),yPos(y),creatureID(ID){setPos(x,y); setTextureAndSprite(ID);}
 
-    bool isHit(int x, int y)
+    bool isHit(float x, float y)
     {
         if (alive && x > xPos && x < xPos + xDim && y > yPos && y < yPos + yDim)
         {
@@ -46,11 +47,11 @@ public:
     }
 
     // Setters
-    void setXDim(int val){xDim = val;}
-    void setYDim(int val){yDim = val;}
-    void setXPos(int val){xPos = val; spriteCreature.setPosition(xPos,yPos);}
-    void setYPos(int val){yPos = val; spriteCreature.setPosition(xPos,yPos);}
-    void setPos(int x, int y){xPos = x; yPos = y; spriteCreature.setPosition(x,y);}
+    void setXDim(float val){xDim = val;}
+    void setYDim(float val){yDim = val;}
+    void setXPos(float val){xPos = val; spriteCreature.setPosition(xPos,yPos);}
+    void setYPos(float val){yPos = val; spriteCreature.setPosition(xPos,yPos);}
+    void setPos(float x, float y){xPos = x; yPos = y; spriteCreature.setPosition(x,y);}
     void setAngle(float val){angle = val; spriteCreature.setRotation(val);}
     void setAlive(bool val){alive = val;}
     void setCreatureID(std::string name){creatureID = name;}
@@ -63,14 +64,14 @@ public:
     void setToken(ScoreToken val){token = val;}
 
     // Getters
-    int getXDim(){return xDim;}
-    int getYDim(){return yDim;}
-    int getXBoundMin(){return xPos;}
-    int getXBoundMax(){return xPos + xDim;}
-    int getYBoundMin(){return yPos;}
-    int getYBoundMax(){return yPos + yDim;}
-    int getXPos(){return xPos;}
-    int getYPos(){return yPos;}
+    float getXDim(){return xDim;}
+    float getYDim(){return yDim;}
+    float getXBoundMin(){return xPos;}
+    float getXBoundMax(){return xPos + xDim;}
+    float getYBoundMin(){return yPos;}
+    float getYBoundMax(){return yPos + yDim;}
+    float getXPos(){return xPos;}
+    float getYPos(){return yPos;}
     float getAngle(){return angle;}
     bool getAlive(){return alive;}
     std::string getCreatureID(){return creatureID;}
@@ -79,10 +80,10 @@ public:
     ScoreToken getToken(){return token;}
 
 private:
-    int xDim;
-    int yDim;
-    int xPos;
-    int yPos;
+    float xDim;
+    float yDim;
+    float xPos;
+    float yPos;
     float angle;
     bool alive;
     std::string creatureID;
